hotel_api/requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy[asyncio]
asyncmy
passlib>=1.7.4
bcrypt==3.2.0
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings
python-dotenv

python-jose[cryptography]
alembic
requests